Inline DropTableTask into DropTablePlan
